Move gdal_utils to root
